pyproject.toml
[project]
name = "noise-becomes-image"
version = "0.1.0"
description = "Particle-based image reconstruction visualization using score-based diffusion"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "imageio>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
